+++ app/Http/Controllers/TypeController.php
@@ -14,7 +14,8 @@ class TypeController extends Controller
         $this->middleware('admin');
     }
 
-    public function index(){
+    public function index()
+    {
         $types = Type::all();
         return view('types.index', compact('types'));
     }
@@ -46,6 +47,10 @@ class TypeController extends Controller
     public function edit($id)
     {
         $type = Type::findOrFail($id);
+        if(Auth::user()->role != User::ADMIN && $type->creator_id != Auth::user()->id){
+            notify()->flash('You dont have permission to edit this type', 'warning');
+            return redirect('types');
+        }
         return view('types.edit', compact('type'));
     }
 
@@ -59,6 +64,10 @@ class TypeController extends Controller
 
         $input = $request->all();
         $type = Type::findOrFail($id);
+        if(Auth::user()->role != User::ADMIN && $type->creator_id != Auth::user()->id){
+            notify()->flash('You dont have permission to edit this type', 'warning');
+            return redirect('types');
+        }
 
         $type->update($input);
 
@@ -70,6 +79,10 @@ class TypeController extends Controller
     public function destroy($id)
     {
         $type = Type::findOrFail($id);
+        if(Auth::user()->role != User::ADMIN && $type->creator_id != Auth::user()->id){
+            notify()->flash('You dont have permission to delete this type', 'warning');
+            return redirect('types');
+        }
         $type->delete();
         notify()->flash('You have successfully deleted a type', 'success');
         return redirect('types');
